Transform task
- Using Saxon6 for processing XSLT1.0 stylesheets
- Basic configuration:
    - input file
    - stylesheet
    - resolver and parser classes
    - verbosity
    - url resolution
- Can register a URLStreamHandler for resolving `classpath:` URLs

The last one is useful for letting a customization layer in one package
refer to stylesheets in a standard distribution in a different package.
One way to do this is to refer to the public URL within the custom
stylesheet, and then rewrite the URL base to a `classpath:` URL in a XML
catalog file.

For example:

```
  <rewriteURI
    uriStartString="http://docbook.sourceforge.net/release/xsl/current/"
    rewritePrefix="classpath:docbook/"
  />
```

would allow this import from a stylesheet in 'my-customization-layer-1.0.jar'

```
  <xsl:import href="http://docbook.sourceforge.net/release/xsl/current/xhtml5/docbook.xsl" />
```

to be resolved to a stylesheet in the standard DocBook distribution in
'net.sf.docbook:docbook-xsl:1.79.1:ns-resources@zip'. The
`classpath:docbook/xhtml5/docbook.xsl` URL is resolved same as if we had done
`getClass().getClassLoader().getResource("docbook/xhtml5/docbook.xsl")`.